+++ torchmd/forcefield.py
@@ -48,11 +48,11 @@ class Forcefield:
 
     def make_lj(self,atomtype_map):
         sorted_keys = [x[0] for x in sorted(atomtype_map.items(), key=lambda item: item[1])]
-        sigma = np.array([self.ff["lj"][at]["sigma"] for at in sorted_keys], dtype=np.float32)
-        epsilon = np.array([self.ff["lj"][at]["epsilon"] for at in sorted_keys], dtype=np.float32)
+        sigma = np.array([self.ff["lj"][at]["sigma"] for at in sorted_keys], dtype=np.float64)
+        epsilon = np.array([self.ff["lj"][at]["epsilon"] for at in sorted_keys], dtype=np.float64)
         A, B = calculateAB(sigma, epsilon)
-        A = torch.tensor(A)
-        B = torch.tensor(B)
+        A = torch.tensor(A).double()
+        B = torch.tensor(B).double()
         return A,B
         
         
@@ -75,7 +75,7 @@ class Forcefield:
                     f"{pair_atomtype} doesn't have bond information in the FF"
                 )
             bond_params.append([bp["k0"], bp["req"]])
-        bond_params = torch.tensor(bond_params)
+        bond_params = torch.tensor(bond_params).double()
         bonds = torch.tensor(uqbonds)
         return bond_params,bonds
 
@@ -100,7 +100,7 @@ class Forcefield:
                     f"{pair_atomtype} doesn't have bond information in the FF"
                 )
             angle_params.append([ap["k0"], np.deg2rad(ap["theta0"])])
-        angle_params = torch.tensor(angle_params)
+        angle_params = torch.tensor(angle_params).double()
         angles = torch.tensor(uqangles)
         return angle_params, angles
